udemy_protobuf_dotnet/pyproject.toml
[project]
name = "udemy-protobuf-dotnet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14,<3.15"
dependencies = [
    "grpcio",
    "grpcio-tools",
    "protobuf",
    "faker",
    "pytest",
    "ipython",
    "ipdb",
    "invoke",
    "watchfiles",
    "icecream>=2.1.10",
    "loguru>=0.7.3",
    "watchdog>=6.0.0",
    "grpcio-reflection>=1.78.1",
    "aioconsole>=0.8.2",
    "ruff>=0.15.5",
]
